Move tests for `mf_pformat_dict` to a separate module.
